doc: update s5 document
For WaaG VM, the User VM name is hard coded in the
lifecycle manager, this User VM ("windows") is needed
for guest shutdown.
Note: libvirt will be used to do guest shutdown and
the WaaG VM name will be configurable in furture.

The lifecycle manager starup picture in WaaG VM is out
of data, update it in this patch.

Tracked-on: #6652
Signed-off-by: Xiangyang Wu <xiangyang.wu@intel.com>