board_inspector: add index info to INTx pins from device objects
INTx routing can be specified in ACPI PRT (PCI Routing Table) in two ways:

    1. By an absolute interrupt line from the global interrupt pool.
    2. By a device object and an index to its interrupt resource.

The board inspector today only encodes the source object of an interrupt
pin routing in XML device resource descriptors, which is incomplete if the
source is a device object (i.e. case 2 above).

This patch adds the index from PRT to device resource nodes as well.

Tracked-On: #7058
Signed-off-by: Junjie Mao <junjie.mao@intel.com>
Acked-by: Anthony Xu <anthony.xu@intel.com>